v0.4.0

- Переход к классическому семантическому версионированию

- Реализованы стратегии WinNT и Unix для заданий 
  "Диспетчеризация процессов"

- Добавлен пункт меню "Справка" с выводом краткого описания проекта

- Добавлено описание структуры файла заданий

- Улучшения и исправления в UI и генераторе заданий

- Обновлены версии сторонних библиотек

- Сгенерированные задания сохраняются во временную папку пользователя

- Добавлен логотип приложения (см. вложения)

- Автоматические сборки в Travis CI и AppVeyor CI отключены из-за 
  ошибок CI-окружения при сборке проекта

- Целевая плтаформа - Windows 10 x64; бинарные файлы будут собираться 
  только для нее, тестирование будет проводится только на этой 
  платформе. В качестве опции обеспечена возможность сборки проекта в 
  macOS 10.14 и Ubuntu 18.04. Исправлены специфичные для macOS ошибки

- Сборка проекта с помощью MinGW больше не поддерживается из-за проблем 
  с производительностью компилятора

- Библиотеки Qt и runtime теперь компонуются динамически. Программа 
  поставляется в виде архива с исполняемым файлом и необходимыми 
  библиотеками

+++ gui/widgets/processestablewidget.cpp
@@ -43,8 +43,13 @@ ProcessesTableWidget::ProcessesTableWidget(QWidget *parent)
   setSelectionMode(QAbstractItemView::SingleSelection);
   setEditTriggers({QTableWidget::NoEditTriggers});
 
-  setHorizontalHeaderLabels({"PID"_qs, "PPID"_qs, "Состояние"_qs,
-                             "Приториет"_qs, "Б. приоритет"_qs, "t"_qs, "T"_qs,
+  setHorizontalHeaderLabels({"PID"_qs,
+                             "PPID"_qs,
+                             "Состояние"_qs,
+                             "Приториет"_qs,
+                             "Б. приоритет"_qs,
+                             "t"_qs,
+                             "T"_qs,
                              "T-t"_qs});
   setSortingEnabled(true);
   resizeColumnsToContents();
@@ -66,8 +71,8 @@ void ProcessesTableWidget::setProcesses(const ProcessesList &processes) {
     setItem(row, 0, new NumberItem(p.pid()));
     setItem(row, 1, new NumberItem(p.ppid()));
     setItem(row, 2, new TextItem(stateMap[p.state()]));
-    setItem(row, 3, new NumberItem(p.priority()));
-    setItem(row, 4, new NumberItem(p.basePriority()));
+    setItem(row, 3, new NumberItem(static_cast<int32_t>(p.priority())));
+    setItem(row, 4, new NumberItem(static_cast<int32_t>(p.basePriority())));
     setItem(row, 5, new NumberItem(p.timer()));
     setItem(row, 6, new NumberItem(p.workTime()));
     setItem(row, 7, new NumberItem(p.workTime() - p.timer()));
